first upload of files
my first small project after delving into a basic javascript course about variables, functions and other stuff. It's a simple custom sign calculator where the User input what they want their sign to say and it calculates the 'price' it would cost them. They can also change the price for each tile and for shipping and see the figures change.

Some issues with it already which are stated on the page itself, will fix as i go but want to document it.